requirements.txt
discord.py
langchain
matplotlib
numpy
openai
pandas
python-dotenv
